services/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "amplify-api"
version = "0.1.0"
description = "Amplify API service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "arq>=0.25.0",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "faster-whisper>=1.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "httpx>=0.26.0", "pytest-asyncio>=0.23.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
